requirements.txt
# Streamlit for the app
streamlit>=1.31

# Environment variable support
python-dotenv>=1.0

# LangChain core and community packages
langchain==1.2.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
langchain-text-splitters>=0.3.0

# PDF reading
PyPDF2>=3.0.1

# Vector store
faiss-cpu>=1.7.4

# Hugging Face embeddings
sentence-transformers>=2.6.0
